Move the some methods and class in rich message pod for acessing
